% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas}
\alias{atlas}
\title{Cuvier's gazelle pedigree}
\format{
A data frame with with records from 948 individuals (in rows), and 10 variables:
\itemize{
  \item{\emph{id} - } Individual identity.
  \item{\emph{dam} - } Maternal identity.
  \item{\emph{sire} - } Paternal identity.
  \item{\emph{survival15} - } 15-days survival.
  \item{\emph{prod} - } Female productivity.
  \item{\emph{sex} - } Individual sex.
  \item{\emph{yob} - } Year of birth.
  \item{\emph{pom} - } Period of management.
  \item{\emph{target} - } Individual in the target population.
  \item{\emph{eeza_id} - } Individual identity (as recorded in the original studbook)
}
}
\source{
The original studbook containing the complete and updated pedigree can be found at: \url{http://www.eeza.csic.es/en/programadecria.aspx}.
}
\usage{
atlas
}
\description{
This data set contains the pedigree of Cuvier's gazelle (\emph{Gazella atlas}).
A total of 948 individuals is included, as well as measurements of biological fitness
and other factors (see reference below for details).
}
\references{
\itemize{
  \item{López-Cortegano E et al. 2021. Genetic purging in captive endangered ungulates with extremely low effective population sizes. *Pending revision*.}
}
}
\keyword{datasets}
