% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_ancestors}
\alias{check_ancestors}
\title{Check ancestor individuals}
\usage{
check_ancestors(id, ancestors)
}
\arguments{
\item{id}{Vector of individual ids.}

\item{ancestors}{Vector of ancestor ids.}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Takes a column name, and checks its use as target.
It should name a boolean vector (or coercible to it),
with at least one TRUE value.
}
