% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_basic}
\alias{check_basic}
\title{Check basic}
\usage{
check_basic(
  ped,
  id_name = "id",
  dam_name = "dam",
  sire_name = "sire",
  when_rename = FALSE,
  when_sort = FALSE
)
}
\arguments{
\item{ped}{A dataframe containing the pedigree. Individual (id), maternal (dam), and paternal (sire) identities are mandatory columns.}

\item{id_name}{Column name for individual id.}

\item{dam_name}{Column name for dam.}

\item{sire_name}{Column name for sire.}

\item{when_rename}{True when called from ped_rename function. It softs checks on individual ID column name and types}

\item{when_sort}{True when called from ped_sort function. It softs checks on pedigree sorting}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
This function will group some other checking functions, that should be run everytime when using
functions in this package, to avoid unexpected errors.
}
