% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_bool}
\alias{check_bool}
\title{Check if a variable is boolean or not}
\usage{
check_bool(variable)
}
\arguments{
\item{variable}{Variable to test}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Can be used to test arguments that need to be of logical (boolean) type
}
