% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_col}
\alias{check_col}
\title{Check that optional column is included}
\usage{
check_col(names, name)
}
\arguments{
\item{names}{Column names (all)}

\item{name}{Column name to check.}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Some functions require additional columns. Check that they are named in the pedigree.
}
