% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_d}
\alias{check_d}
\title{Check purging coefficient}
\usage{
check_d(d)
}
\arguments{
\item{d}{Purging coefficient (taking values between 0.0 and 0.5).}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
The purging coefficient must be a number between 0 and 0.5
}
