% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_df}
\alias{check_df}
\title{Check pedigree class}
\usage{
check_df(obj)
}
\arguments{
\item{obj}{Object to test}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
The pedigree must be of object class 'data.frame'.
}
