% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_Fcol}
\alias{check_Fcol}
\title{Check columns with inbreeding values}
\usage{
check_Fcol(ped, Fcol, compute = TRUE)
}
\arguments{
\item{ped}{A dataframe containing the pedigree. Individual (id), maternal (dam), and paternal (sire) identities are mandatory columns.}

\item{Fcol}{Name of column with inbreeding coefficient values. If none is used, inbreeding will be computed.}

\item{compute}{Compute inbreeding if Fcol is NULL}
}
\value{
Vector of inbreeding values (if checks are successful)
}
\description{
Takes a column name, and checks its use as inbreeding coefficient.
It should name a numeric vector, with values in the range [0,1]
}
