% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_index}
\alias{check_index}
\title{Check individual index}
\usage{
check_index(id)
}
\arguments{
\item{id}{Column of individual ids.}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Renamed individuals must be named by their index (from 1 to N)
}
