% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_int}
\alias{check_int}
\title{Check if a variable is a positive integer or not}
\usage{
check_int(variable)
}
\arguments{
\item{variable}{Variable to test}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Can be used to test arguments that need to be integers
}
