% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_length}
\alias{check_length}
\title{Check if a variable has length >1}
\usage{
check_length(variable, message = "Expected value of length 1")
}
\arguments{
\item{variable}{Variable to test}

\item{message}{Error message to display}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Used to test arguments that need to be of length 1
}
