% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_na}
\alias{check_na}
\title{Check if a vector contains NA values}
\usage{
check_na(variable)
}
\arguments{
\item{variable}{Variable to test}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Return warning when NA values are present
}
