% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_names}
\alias{check_names}
\title{Check that mandatory column names are included}
\usage{
check_names(ped, id_name = "id", dam_name = "dam", sire_name = "sire")
}
\arguments{
\item{ped}{A dataframe containing the pedigree. Individual (id), maternal (dam), and paternal (sire) identities are mandatory columns.}

\item{id_name}{Column name for individual id.}

\item{dam_name}{Column name for dam.}

\item{sire_name}{Column name for sire.}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Columns for id, dam and sire are mandatory. This function checks that they are named in the pedigree.
The function works with arbitrary column names (not 'id', 'dam' and 'sire') to work with ped_rename()
}
