% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_Ne}
\alias{check_Ne}
\title{Check Ne}
\usage{
check_Ne(Ne)
}
\arguments{
\item{Ne}{Effective population size}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
The effective population size (Ne) must be a number higher than 0
}
