% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_order}
\alias{check_order}
\title{Check individual order}
\usage{
check_order(id, dam, sire, soft_sorting = FALSE)
}
\arguments{
\item{id}{Vector of individual ids.}

\item{dam}{Vector of dam ids.}

\item{sire}{Vector of sire ids.}

\item{soft_sorting}{If TRUE checking is relaxed, allowing descendants to be declared before ancestors}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Individuals must be sorted from older to younger
}
