% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_repeat_id}
\alias{check_repeat_id}
\title{Check repeated ids}
\usage{
check_repeat_id(id)
}
\arguments{
\item{id}{Vector of individual ids.}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Individual id are unique and cannot be repeated
}
