% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_tcol}
\alias{check_tcol}
\title{Check columns with generation numbers}
\usage{
check_tcol(ped, tcol, compute = TRUE, force_int = FALSE)
}
\arguments{
\item{ped}{A dataframe containing the pedigree. Individual (id), maternal (dam), and paternal (sire) identities are mandatory columns.}

\item{tcol}{Name of column with individual generation times. If none is used, the number of equivalent complete generations is computed.}

\item{compute}{Compute generation numbers if tcol is NULL}

\item{force_int}{Generation numbers must be integers (disabled by default)}
}
\value{
Vector of generation numbers (if checks are successful)
}
\description{
Takes a column name, and checks its use as generation numbers.
It should name a numeric vector, with values >= 0.
}
