% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_types}
\alias{check_types}
\title{Check that mandatory column names are of type int}
\usage{
check_types(id, dam, sire)
}
\arguments{
\item{id}{Vector of individual ids.}

\item{dam}{Vector of dam ids.}

\item{sire}{Vector of sire ids.}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Columns for id, dam and sire are mandatory, and required to be of type integer
}
