% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_zero_id}
\alias{check_zero_id}
\title{Check individuals named zero}
\usage{
check_zero_id(id)
}
\arguments{
\item{id}{Vector of individual ids.}
}
\value{
No return value. Will print an error message if checking fail.
}
\description{
Individual id cannot equal zero (0). This is reserved to dams and sires.
}
