% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darwin}
\alias{darwin}
\title{Darwin/Wedgwood pedigree}
\format{
A data frame with with records from 63 individuals (in rows), and 3 variables:
\itemize{
  \item{\emph{Individual} - } Individual identity.
  \item{\emph{Mother} - } Mother's identity.
  \item{\emph{Father} - } Father's identity.
}
}
\source{
The pedigree is adapted from Berra et al. (2010)
}
\usage{
darwin
}
\description{
This data set contains the pedigree of the Darwin/Wedgwood dynasty.
It is composed by a total of 63 individuals,
including Charles R. Darwin and Francis Galton.
}
\references{
\itemize{
  \item{Berra TM et al. 2010. Was the Darwin/Wedgwood dynasty adversely affected by consanguinity?. BioScience 60(5): 376-383.}
}
}
\keyword{datasets}
