% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ne.R
\name{delta_Fi}
\alias{delta_Fi}
\title{Individual inbreeding variation}
\usage{
delta_Fi(Fi, t)
}
\arguments{
\item{Fi}{Individual inbreeding coefficient.}

\item{t}{Individual generation number.}
}
\value{
Individual variation in inbreeding.
}
\description{
Computes the increase in inbreeding coefficient for a given individual
}
