% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate}
\alias{evaluate}
\title{Individuals to be evaluated in purging analyses}
\usage{
evaluate(ped, value_from)
}
\arguments{
\item{ped}{A dataframe containing the pedigree. Individual (id), maternal (dam), and paternal (sire) identities are mandatory columns.}

\item{value_from}{Name of the column of interest.}
}
\value{
Boolean vector indicating what individuals will be evaluated.
}
\description{
Returns a boolean vector indicating what individuals are suitable for purging analyses, given a measure of fitness.
Individuals with NA values of fitness, and that do not have descendants with non-NA fitness values, are excluded.
}
