% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{idx_ancestors}
\alias{idx_ancestors}
\title{Index ancestors}
\usage{
idx_ancestors(ids, N)
}
\arguments{
\item{ids}{Ancestor identities}

\item{N}{Total number of individuals}
}
\value{
A logical matrix.
}
\description{
Creates a vector of length N (the number of individuals)
Only coordinates for valid ancestors will be given
}
