% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_allele}
\alias{sample_allele}
\title{Sample dam or sire inherited allele}
\arguments{
\item{dam_al}{Dam allele.}

\item{sire_al}{Sire allele.}
}
\value{
The sampled allele.
}
\description{
Given two alleles (one from dam, the other from sire), it samples one at random.
}
