% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search_ancestors}
\alias{search_ancestors}
\title{Search and individuals' ancestors}
\arguments{
\item{dam}{Vector of dams.}

\item{sire}{Vector of sires.}

\item{i}{Reference individual (its index, not id).}

\item{fnd}{Vector of founders (to be returned as reference).}

\item{anc}{Vector of ancestors (to be returned as reference).}
}
\value{
The sampled allele.
}
\description{
Recursive function that gathers all founders and ancestors for a given individual
}
