% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{w_grandoffspring}
\alias{w_grandoffspring}
\title{Grandoffspring}
\usage{
w_grandoffspring(ped, name_to)
}
\arguments{
\item{ped}{A dataframe containing the pedigree. Individual (id), maternal (dam), and paternal (sire) identities are mandatory columns.}

\item{name_to}{A string naming the new output column.}
}
\value{
The input dataframe, plus an additional column indicating the total number of grandoffspring.
}
\description{
Counts the number of grandoffspring for individuals in the pedigree.
}
\examples{
data(dama)
w_grandoffspring(dama, name_to = "G")
}
