% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Fij_core_i_cpp}
\alias{Fij_core_i_cpp}
\title{Partial inbreeding coefficient (core function)}
\usage{
Fij_core_i_cpp(dam, sire, anc_idx, mapa, Fi, genedrop = 0L, seed = NULL)
}
\arguments{
\item{dam}{Vector of dam ids.}

\item{sire}{Vector of sire ids.}

\item{anc_idx}{Index of ancestors.}

\item{mapa}{Map of ancestors}

\item{Fi}{Vector of inbreeding coefficients.}

\item{genedrop}{Enable genedrop simulation}

\item{seed}{Sets a seed for the random number generator.}
}
\value{
A matrix of partial inbreeding coefficients. Fi(j) values can thus be read from row i and column j.
}
\description{
Computes partial inbreeding coefficients, Fi(j).
A coefficient Fi(j) can be read as the probability of individual i being
homozygous for alleles derived from ancestor j
}
